Use python poetry
This completely converts this project to use poetry
for building and running tests.  The motivation for
using poetry is to make dependency management and
release management more simple.  The only thing
that does not use poetry is deployment to pypi which
is still using twine.  We'll tackle that in another
change.